ajustes de importação tarot
