+++ src/cards.html
@@ -24,7 +24,7 @@
       <nav class="filtro01">
         <ul class="lista">
           <li><a href="index.html" class="home-page">Pagina Inicial</a></li>
-          <li><a href="cards.html?type=maior?type=menor" class="cardall">Todas as Cartas</a></li>
+          <li><a href="cards.html?type=maior?type=menor" class="cardall">Baralho</a></li>
           <li><a href="cards.html?type=maior" class="bigger">Arcanos maiores</a></li>
           <li><a href="cards.html?type=menor" class="minors">Arcanos menores</a></li>
           <li><a href="cards.html?suit=copas" class="copas">Baralho de copas</a></li>
