--- src/data.js
@@ -1,9 +1,0 @@
-// estas funciones son de ejemplo
-
-export const tarot = () => {
-  return 'tarot';
-};
-
-export const anotherTarot = () => {
-  return 'OMG';
-};
