+++ src/index.html
@@ -27,7 +27,7 @@
           <li><a href="cards.html?type=maior?type=menor" class="cardall">Todas as Cartas</a></li>
           <li><a href="cards.html?type=maior" class="bigger">Arcanos maiores</a></li>
           <li><a href="cards.html?type=menor" class="minors">Arcanos menores</a></li>
-          <li><a href="cards.html?suit=copas" class="copas">Baralho de copas</a></li>
+          <li><a href="cards.html?suit=copas" class="copas active">Baralho de copas</a></li>
           <li><a href="cards.html?suit=ouros" class="ouros">Baralho de Ouros</a></li>
           <li><a href="cards.html?suit=paus" class="paus">Baralho de Paus</a></li>
           <li><a href="cards.html?suit=espadas" class="espadas">Baralho de Espadas</a></li>
